PR findings
